requirements.txt
aws-cdk-lib==2.171.1
aws-cdk.lambda-layer-kubectl-v31
constructs>=10.0.0
boto3>=1.28.15
botocore>=1.31.15
bandit>=1.7.0